#18 feat: Update cropper.js for image cropping functionality and improve team icon editing page
