+++ src/todo-gantt/resources/css/app.css
@@ -1,3 +1,5 @@
+@import 'cropperjs/dist/cropper.css';
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
